Обновил ресурсы и добавление ЛИР, ВЛИР, и так по мелочи

+++ web/SandBox.WebUi/Pages/Research/Current.aspx.cs
@@ -216,7 +216,7 @@ namespace SandBox.WebUi.Pages.Research
 
                 ResearchManager.UpdateResearchState(researchId, ResearchState.EXECUTING);
                 ResearchManager.UpdateResearchStartTime(researchId); //?? Должно быть выше
-                VmManager.UpdateVmState(vm.Id, VmManager.State.RESEARCHING);
+                VmManager.UpdateVmState(vm.Id, (Int32)VmManager.State.RESEARCHING);
                 ResearchManager.UpdateResearchVmData(ResearchManager.GetResearch(researchId).ResearchVmData, vm.EnvMac, vm.EnvIp);
             }
             else
